+++ packages/auto_animated/lib/auto_animated.dart
@@ -1,7 +1,8 @@
+export 'src/animate_if_visible.dart';
 export 'src/grid.dart';
 export 'src/helpers/callbacks.dart';
+export 'src/helpers/options.dart';
 export 'src/icon_button.dart';
 export 'src/list.dart';
-export 'src/on_visibility_change.dart';
 export 'src/sliver_grid.dart';
 export 'src/sliver_list.dart';