+++ packages/native_color/lib/native_color.dart
@@ -1,4 +1,5 @@
 export 'src/cielab.dart';
+export 'src/helpers.dart';
 export 'src/hex.dart';
 export 'src/hsl.dart';
 export 'src/xyz.dart';