+++ packages/auto_animated/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.0.1 
+
+* Fix missing import for `LiveOptions` ([issue#23](https://github.com/rbcprolabs/packages.flutter/issues/23))
+
+## 2.0.0
+
+* All widgets with pattern `AutoAnimated{List,Grid,SliverList,SliverGrid,IconButton}` renamed to `Live{List,Grid,SliverList,SliverGrid,IconButton}`. For example: `AutoAnimatedList` now named `LiveList`
+* `AnimateOnVisibilityWrapper` renamed to `AnimateIfVisibleWrapper` and `AnimateOnVisibilityChange` renamed to `AnimateIfVisible`
+* Added `{List,Grid,SliverList,SliverGrid,IconButton}.options` constructors
+* Removed `AutoAnimated.separated()` constructor. Now property `separatorBuilder` added in `LiveList()` & `LiveList.options()` constructor and marked as optional
+* Added `LiveOptions`
+* Added `visibleFraction` as option for `Live{List,Grid,SliverList,SliverGrid,IconButton}` & `LiveOptions`. Animations starts at `double visibleFraction` visible item fraction in sight (default 0.025)
+* Docs improvements
+
 ## 2.0.0-dev.4
 
 * SliverList & SliverGrid now requires scrollController (From custom `CustomScrollView`). It is necessary in order to find out in which direction the widget scrolls in order to play the animation in the corresponding direction
@@ -66,7 +80,7 @@ Now animations start on scroll (element visibility change)
 
 ## 1.1.0
 
-* Added `AutoAnimatedListState.sepparated`
+* Added `AutoAnimatedListState.separated`
 * Added `AutoAnimatedIconButton`
 * Optimized performance