+++ packages/native_color/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.2.0
+
+* Added `Color.{lighter(int percents), .darker(int percents)}` extension method helpers
+* Set Dart SDK to 2.6.0
+
+## 1.1.0
+
+* Added source hex string field `hexColor` to `HexColor`
+
 ## 1.0.0
 
 * Added examples